Android: increase shader precision to avoid bleeding problems

+++ cocos/renderer/shaders/positionTextureColor.vert
@@ -31,8 +31,8 @@ attribute vec4 a_color;
 uniform mat4 u_MVPMatrix;
 
 #ifdef GL_ES
-varying lowp vec4 v_fragmentColor;
-varying mediump vec2 v_texCoord;
+varying highp vec4 v_fragmentColor;
+varying highp vec2 v_texCoord;
 #else
 varying vec4 v_fragmentColor;
 varying vec2 v_texCoord;
